DTP (Test > Production: 2bd5e4e1)

+++ dashboard/app/models/programming_expression.rb
@@ -264,6 +264,7 @@ class ProgrammingExpression < ApplicationRecord
   end
 
   def remove_serialization
+    return unless Rails.application.config.levelbuilder_mode
     File.delete(file_path) if File.exist?(file_path)
   end
